Fixing streamer build and installation.

+++ src/modules/stream/install_required.py
@@ -1,6 +1,9 @@
 import os, sys, apt, platform, subprocess, re
 
 def replace_regex_in_file_line(filename, original, replacement, dry_run = False):
+    if dry_run:
+        print "Replacing", original, "with", replacement, "in", filename
+        return
     with open(filename, "r") as input:
         lines = input.readlines()
     with open(filename, "w") as output:
@@ -16,18 +19,23 @@ def execute(shell_cmd_string, directory = "", dry_run = False):
         print "Executing", shell_cmd_string, "in", cwd
 
 
-def apt_get(package_name, clean = False, dry_run = False):
-    apt_cache = apt.Cache()
+def apt_get(package_name, clean = False, dry_run = False, prefix = None):
+    apt_cache = apt.cache.Cache()
+    # apt_cache = apt.Cache() # no difference
+    print "Updating apt-cache..."
+    apt_cache.update()
     try:
         package = apt_cache[package_name]
         if not package.is_installed:
-            if dry_run:
-                print "Installing system package " + package_name
-            else:
+            print "Installing system package " + package_name
+            if not dry_run:
                 package.mark_install()
                 apt_cache.commit()
         else:
             print "Skipping package " + package_name + ", which is already installed"
+        print "Done"
+    except TypeError, te:
+        print "Wrong types: " + str(e)
     except KeyError, ke:
         print "Unknown package " + package_name
     except apt.cache.LockFailedException, lfe:
@@ -36,10 +44,11 @@ def apt_get(package_name, clean = False, dry_run = False):
 
 # install header and dynamic libraries for the live555 video stremer
 def live(address, clean, dry_run, prefix = "/usr/local/"):
+    print "Installing live555 in " + prefix
     dirs = ['BasicUsageEnvironment', 'UsageEnvironment', 'groupsock', 'liveMedia']
 
     if clean:
-        execute('sudo rm -rf liveMedia', prefix + 'include', dry_run)
+        execute('sudo rm -rf ' + ' '.join(dirs), prefix + 'include', dry_run)
         for d in dirs:
             execute('sudo rm -rf lib' + d + '*', prefix + 'lib/', dry_run)
         if os.path.isdir("/tmp/live555"):
@@ -57,17 +66,18 @@ def live(address, clean, dry_run, prefix = "/usr/local/"):
     execute('./genMakefiles ' + target, '/tmp/live555', dry_run)
 
     # configure for install in /usr instead of /usr/local ??
-    if prefix != "/usr/":
+    if prefix != "/usr/local":
         for dir in dirs:
             replace_regex_in_file_line('/tmp/live555/' + dir + '/Makefile',
-                                       prefix, '/usr', dry_run)
+                                       '/usr/local/', prefix, dry_run)
 
     execute('make -j4', '/tmp/live555', dry_run)
 
     for dir in dirs:
+        print 'sudo make install', '/tmp/live555/' + dir
         execute('sudo make install', '/tmp/live555/' + dir, dry_run)
 
-    print "--> Live555 installed, run: sudo ldconfig " + prefix + "lib/"
+    print "--> Live555 installed"
 
 if __name__ == '__main__':
 
@@ -84,9 +94,9 @@ if __name__ == '__main__':
             clean = True
 
     requirements = {
-#        'libx264-dev' : apt_get,
+        'libx264-dev' : apt_get,
         'https://github.com/hackeron/live555.git' : live
     }
 
     for req in requirements:
-        requirements[req](req, clean, dry_run)
+        requirements[req](req, clean, dry_run, prefix = "/usr/")
